+++ packages/client-node/src/connection/node_custom_agent_connection.ts
@@ -26,8 +26,8 @@ export class NodeCustomAgentConnection extends NodeBaseConnection {
   protected createClientRequest(params: RequestParams): Http.ClientRequest {
     const headers = withCompressionHeaders({
       headers: params.headers,
-      enable_request_compression: params.enable_request_compression,
-      enable_response_compression: params.enable_response_compression,
+      request_compression_codec: params.request_compression_codec,
+      response_compression_codec: params.response_compression_codec,
     });
     return this.httpRequestFn(params.url, {
       method: params.method,